Mark GUIControl destructor as virtual
Previously GUIControl destructor was not virtual and did not run any of
derived classes destructors. Now that it is and derived destructors are
ran, they actually double free due to `Destroy` has already been called
on the objects by `GUIControlManager::Clear`. They were also
inconsistent with some classes having them and some not, so I unified
the approach and removed the destructors in `GUIControl`-derived classes
that only called `Destroy`.