Improve LimbPath.cpp - rewrite ReportProgress, etc
The rewrite checks all path segments from the current one to the last and finds
the best one to continue from, instead of fixating on just the one. This helps
in cases where feet might be yanked along the path by excessive horizontal
acceleration and removes quite a bunch of stumbling.

Also fix Vector/Matrix division, making it the proper inverse of multiplication
and applicable to const matrix.